requirements.txt
pandas==1.5.3
python-dotenv==0.21.0
scikit-learn==1.2.0
matplotlib==3.6.3
numpy==1.20.3
streamlit==1.10.0
streamlit-folium==0.6.15
seaborn==0.12.1
langchain 
faiss-cpu
fastapi
python-dotenv
langchain
uvicorn
openai
tiktoken

